Setup terrain as a module
Terrain class moved to terrain_handler.py, and no longer singleton
Instance of Terrain class is created in __init__.py
