+++ game/objects/player.py
@@ -13,8 +13,6 @@ class Player(pyglet.sprite.Sprite):
     def __init__(self, *args, **kwargs):
         super(Player, self).__init__(img=resources.player_image, *args, **kwargs)
 
-        self.terrain = terrain.Terrain()
-
         self.register_event_type("on_move")
 
         self.key_handler = key.KeyStateHandler()
@@ -92,7 +90,7 @@ class Player(pyglet.sprite.Sprite):
     def calculate_move_xory(self, dpos, speed):
         newpos = self.pos + dpos * (self.size // 2 + speed)
 
-        tile = self.terrain.get_tile(newpos)
+        tile = terrain.get_tile(newpos)
 
         normal_move = dpos * speed
         snap_move = (abs(tile.screenpos - self.screenpos) - tile.size // 2 - self.size // 2) * dpos
@@ -100,16 +98,16 @@ class Player(pyglet.sprite.Sprite):
         # Check if new pos is obstructed
         if tile.material != "air":
             # Test if we can move UP to the next tile
-            tile_a = self.terrain.get_tile(newpos + positions.Pos3(0, 0, 1))
+            tile_a = terrain.get_tile(newpos + positions.Pos3(0, 0, 1))
             if tile_a.material == "air":
                 return normal_move, 1
 
             return snap_move, 0
         
         # Check if there is a tile below new pos
-        if self.terrain.get_tile(newpos - positions.Pos3(0, 0, 1)).material == "air":
+        if terrain.get_tile(newpos - positions.Pos3(0, 0, 1)).material == "air":
             # Test if we can move DOWN to the next tile
-            if self.terrain.get_tile(newpos - positions.Pos3(0, 0, 2)).material != "air":
+            if terrain.get_tile(newpos - positions.Pos3(0, 0, 2)).material != "air":
                 return normal_move, -1
             
             return snap_move, 0
@@ -127,7 +125,7 @@ class Player(pyglet.sprite.Sprite):
             self.dispatch_event("on_move")
     
     def on_move(self):
-        self.terrain.update(self.pos)
+        terrain.update(self.pos)
 
     @pause.pausable
     def on_mouse_motion(self, x, y, dx, dy):
