+++ game/terrain/__init__.py
@@ -1,0 +1,6 @@
+from .terrain_handler import *
+from .data_handler import *
+from .chunk import *
+from .tile import *
+
+terrain = Terrain()
